% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_profit_sorted}
\alias{plot_profit_sorted}
\title{Bar plot of arbitrage profit for each tick, sorted}
\usage{
plot_profit_sorted(x, separate_plots = FALSE, ...)
}
\arguments{
\item{x}{'TriArbProfitTable' object or a data frame with 4 columns: timestamp,
difference between timestamp, profit of first roundtrip, and profit of second
roundtrip.}

\item{...}{further arguments for the generic 'plot' function.}

\item{separate}{logical. If FALSE then one device window will show two plots of the two possible
triangular arbitrage roundtrips. If TRUE, each plot will be opened in new window.}
}
\description{
This function produces a bar plots of triangular arbitrage profit in pips
for each tick. The bars are arranged on the x-axis according to their timestamp.
}

