% Generated by roxygen2 (4.0.2): do not edit by hand
\name{consec_dups}
\alias{consec_dups}
\title{Find consecutive duplicated quotes}
\usage{
consec_dups(x)
}
\arguments{
\item{x}{A data frame with first column being a timestamp of POSIXct class}
}
\value{
A vector of row indexes of duplicated quotes
}
\description{
Find consecutive (when sorted by timestamp) duplicated quotes.
#'
}
\examples{
data(AUDCAD)
dups <- consec_dups(AUDCAD)
AUDCAD <- AUDCAD[-dups, ] # removes the duplicates
}

